% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_phylo.R
\name{get_structure_name_hook.multiPhylo}
\alias{get_structure_name_hook.multiPhylo}
\title{Get Structure Name Hook for multiPhylo}
\usage{
\method{get_structure_name_hook}{multiPhylo}(structure, ...)
}
\arguments{
\item{structure}{The multiPhylo structure.}

\item{...}{Additional arguments.}
}
\description{
Get Structure Name Hook for multiPhylo
}
