% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_phylo.R
\name{jags_structure_definition.phylo}
\alias{jags_structure_definition.phylo}
\title{JAGS Structure Definition for Phylogenetic Trees}
\usage{
\method{jags_structure_definition}{phylo}(
  structure,
  variable_name = "err",
  optimize = TRUE,
  precision_parameter = "lambda",
  ...
)
}
\arguments{
\item{structure}{A \code{phylo} object from the \code{ape} package.}

\item{variable_name}{Name of the error variable (default "err").}

\item{optimize}{Logical. If TRUE (default), uses pre-calculated precision matrix.}

\item{precision_parameter}{Name of the precision parameter (default "lambda").}

\item{...}{Additional arguments (ignored).}
}
\value{
A list with \code{setup_code} and \code{error_prior} JAGS code strings.
}
\description{
Implements the \code{jags_structure_definition} S3 method for \code{phylo} objects.
This method is provided by the \code{because.phybase} package to extend the base
\code{because} engine with phylogenetic covariance support.
}
