% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_phylo.R
\name{numpyro_structure_definition.multiPhylo}
\alias{numpyro_structure_definition.multiPhylo}
\title{Implements the \code{numpyro_structure_definition} S3 method for \code{multiPhylo} objects.}
\usage{
\method{numpyro_structure_definition}{multiPhylo}(structure, variable_name = "err", ...)
}
\arguments{
\item{structure}{The \code{multiPhylo} object.}

\item{variable_name}{The response variable name.}

\item{...}{Additional arguments.}
}
\description{
Implements the \code{numpyro_structure_definition} S3 method for \code{multiPhylo} objects.
}
