% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_phylo.R
\name{numpyro_structure_definition.phylo}
\alias{numpyro_structure_definition.phylo}
\title{NumPyro Structure Definition for Phylogenetic Trees}
\usage{
\method{numpyro_structure_definition}{phylo}(structure, variable_name = "err", ...)
}
\arguments{
\item{structure}{A \code{phylo} object.}

\item{variable_name}{Name of the variable.}

\item{...}{Additional arguments.}
}
\value{
A character string of Python code.
}
\description{
Implements the \code{numpyro_structure_definition} S3 method for \code{phylo} objects.
Returns the JAX/NumPyro Python code to compute the Cholesky factor of the
phylogenetic variance-covariance matrix dynamically scaled by Pagel's lambda.
}
