% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rhino_tree.R
\docType{data}
\name{rhino.tree}
\alias{rhino.tree}
\title{Rhinograd phylogenetic tree}
\format{
A phylo object (from the ape package) with 100 tips and 99 internal nodes.
The tree has been scaled so that the root age is 1.0.
}
\source{
Simulated data for Gonzalez-Voyer and von Hardenberg (2014)
}
\usage{
rhino.tree
}
\description{
A phylogenetic tree (phylo object) for 100 simulated Rhinograd species.
This tree is used to demonstrate phylogenetic Bayesian structural equation
modeling with the becauseR package.
}
\details{
Rhinograds  are hypothetical mammals used as examples in
phylogenetic comparative methods. This simulated tree represents the
evolutionary relationships among 100 Rhinograd species.
}
\examples{
data(rhino.tree)
plot(rhino.tree)

}
\references{
Gonzalez-Voyer, A., & von Hardenberg, A. (2014). An introduction to
phylogenetic path analysis. In: Modern phylogenetic comparative methods
and their application in evolutionary biology (pp. 201–229). Springer.
}
\keyword{datasets}
